добавляет корректировки к заданию
